pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentshield"
version = "0.1.0"
description = "Security evaluation framework for tool-using AI agents and MCP-style workflows"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Samarth Shete" }]
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.2", "ruff>=0.6.2"]

[project.scripts]
agentshield = "agentshield.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentshield*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
